Adventurer Update / Guild DS
Changed the Adventurer class to use a Guild structure for it's parent
guild. Created the Guild structure. I have substantial problems with the
includes, and my compiler is still being an absolute jerk, so I will
need someone else to test the hell out of these and to make sure the
include statements work.